Muevo configs a módulo aparte
