+++ conf/prod.py
@@ -1,0 +1,6 @@
+from conf.local import Config as BaseConfig
+
+
+class Config(BaseConfig):
+    DEBUG = False
+    SKIP_AUTH = False
