+++ config.py
@@ -1,9 +1,1 @@
-import os
-
-import flask
-
-
-class Config(flask.Config):
-    DEBUG = os.environ.get('ENV') == 'development'
-    MONGO_URI = os.environ.get('DB', "mongodb://localhost:27017/development")
-    SKIP_AUTH = False
+from conf.prod import Config  # Backwards comp
